% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottimes-package.R
\docType{package}
\name{dottimes-package}
\alias{dottimes}
\alias{dottimes-package}
\title{dottimes}
\description{
dottimes
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

