% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{times}
\alias{times}
\title{Run an expression N times}
\usage{
times(x, ...)
}
\description{
Run an expression N times
}
\examples{
times(3, print("fart"))
times(3, head(mtcars))
times(10, rnorm(5))
10 \%>\% times(rnorm(5))
}

