% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upto.R
\name{upto}
\alias{upto}
\title{Run an expression from X up to N}
\usage{
upto(min, max, ...)
}
\description{
Run an expression from X up to N
}
\examples{
upto(3, 5, print("fart"))
upto(3, 5, )
upto(2, 4, head(mtcars))
upto(10, 13, rnorm(5))
10 \%>\% upto(15, rnorm(5))

let <- letters[1:10]
upto(1, 6, print(let))
}

